# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trançify SaaS multi-tenant scheduling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login tenant or super admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-email:
    post:
      operationId: changeEmail
      tags: [auth]
      summary: Change email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeEmailRequest"
      responses:
        "200":
          description: Email changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /availability:
    get:
      operationId: getMyAvailability
      tags: [availability]
      summary: Get availability config for current tenant
      responses:
        "200":
          description: Availability config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
    put:
      operationId: updateAvailability
      tags: [availability]
      summary: Update availability config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvailabilityRequest"
      responses:
        "200":
          description: Updated availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
  /availability/public/{tenantId}/dates:
    get:
      operationId: getPublicAvailabilityDates
      tags: [availability]
      summary: Get available dates for a tenant (for calendar display)
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available dates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableDates"
  /availability/public/{tenantId}:
    get:
      operationId: getPublicAvailability
      tags: [availability]
      summary: Get available time slots for a tenant on a given date
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: serviceId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlots"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [token, user]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [tenant, super_admin]
        tenantId:
          type: string
          nullable: true
        tenantSlug:
          type: string
          nullable: true
      required: [id, email, role]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
    ChangeEmailRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newEmail:
          type: string
      required: [currentPassword, newEmail]
...
    Availability:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        availableDays:
          type: array
          items:
            type: integer
            description: Day of week 0=Sun, 1=Mon, ..., 6=Sat
        startTime:
          type: string
          description: HH:MM format
        endTime:
          type: string
          description: HH:MM format
        slotIntervalMinutes:
          type: integer
        breakAfterMinutes:
          type: integer
          description: Rest time after each appointment in minutes
        maxAppointmentsPerDay:
          type: integer
        blockedDates:
          type: array
          items:
            type: string
            format: date
      required: [id, tenantId, availableDays, startTime, endTime, slotIntervalMinutes, breakAfterMinutes, maxAppointmentsPerDay]
    UpdateAvailabilityRequest:
      type: object
      properties:
        availableDays:
          type: array
          items:
            type: integer
        startTime:
          type: string
        endTime:
          type: string
        slotIntervalMinutes:
          type: integer
        breakAfterMinutes:
          type: integer
        maxAppointmentsPerDay:
          type: integer
        blockedDates:
          type: array
          items:
            type: string
            format: date
    AvailableSlots:
      type: object
      properties:
        date:
          type: string
          format: date
        slots:
          type: array
          items:
            type: string
            description: Available time slots in HH:MM format
        available:
          type: boolean
      required: [date, slots, available]
    AvailableDates:
      type: object
      properties:
        availableDates:
          type: array
          items:
            type: string
            format: date
          description: List of available dates in YYYY-MM-DD format
      required: [availableDates]
...