# lib/api-zod/src/generated/types/updateAvailabilityRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trançify SaaS multi-tenant scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAvailabilityRequest {
  availableDays?: number[];
  startTime?: string;
  endTime?: string;
  slotIntervalMinutes?: number;
  breakAfterMinutes?: number;
  maxAppointmentsPerDay?: number;
  blockedDates?: Date[];
}